# Conflicts:
#	src/main/java/appLogic/AppContext.java
#	src/main/java/appLogic/activity/DefaultActivityService.java
#	src/main/java/appLogic/activity/command/CreateWorkActivity.java
#	src/main/java/appLogic/activity/impl/FixedActivity.java
#	src/main/java/appLogic/activity/impl/WorkActivity.java
#	src/main/java/appLogic/project/ProjectActivity.java

+++ src/main/java/appLogic/activity/command/CreateWorkActivity.java
@@ -3,4 +3,4 @@ package appLogic.activity.command;
 import java.time.LocalDateTime;
 
 public record CreateWorkActivity(String projectId, String name, String description, String summary,
-                                 LocalDateTime startDate, LocalDateTime endDate, int expectedHours) {}
+                                 LocalDateTime startDate, LocalDateTime endDate, double expectedHours) {}
